Перенос функций/структур плат в отдельный файл
